% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\name{dataInfo}
\alias{dataInfo}
\title{Display select information about hub object}
\usage{
dataInfo(hub, rdataclass)
}
\arguments{
\item{hub}{A hub object, either AnnotationHub or ExperimentHub.}

\item{rdataclass}{For the hub resources of interest, the class of the R
object used to represent the object when imported into R, e.g., 'GRanges'.}
}
\value{
A DataFrame.
}
\description{
Display select information about hub object
}
\examples{
ah = AnnotationHub()
dataInfo(ah, "BigWigFile")
}
