% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverData.R
\name{discoverData}
\alias{discoverData}
\title{Discover what data source types are available in the specific Hub}
\usage{
discoverData(hub, fileType = types())
}
\arguments{
\item{hub}{A hub object, either AnnotationHub or ExperimentHub.}

\item{fileType}{The type of file the user is intersted in exploring, e.g.
'BigWigFile'.}
}
\value{
A tibble sorted in descending order.
}
\description{
Discover what data source types are available in the specific Hub
}
\examples{
library(AnnotationHub)
ah = AnnotationHub()
discoverData(ah)
}
