% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuery.R
\name{importData}
\alias{importData}
\title{Import data of hub object}
\usage{
importData(hub, ids, which)
}
\arguments{
\item{hub}{A hub object, either AnnotationHub or ExperimentHub}

\item{ids}{The hub id(s) of interest.}

\item{which}{A range data structure, like a 'GRanges'.

importFrom rtracklayer import.bw import.chain import.2bit import.bed
importFrom tools file_ext}
}
\value{
A 'GRanges'.
}
\description{
Import data of hub object
}
\examples{
ah = AnnotationHub()
which <- GRanges(c("chr2", "chr2"), IRanges(c(1, 300), c(400, 1000)))
importData(ah, "AH49544", which)
}
