% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\name{dataInfo}
\alias{dataInfo}
\title{Display select information about hub object}
\usage{
dataInfo(hub, fileType)
}
\arguments{
\item{hub}{A hub object, either AnnotationHub or ExperimentHub.}

\item{fileType}{The type of file the user in interested in exploring, e.g.
'BigWigFile'.}
}
\value{
A DataFrame.
}
\description{
Display select information about hub object
}
\examples{
ah = AnnotationHub()
dataInfo(ah, "BigWigFile")
}
