% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\name{dataInfo}
\alias{dataInfo}
\title{Display select information about hub object}
\usage{
dataInfo(hub, sourceType)
}
\arguments{
\item{hub}{A hub object, either AnnotationHub or ExperimentHub.}

\item{sourceType}{For the hub resources of interest, the format of the
original resource, e.g., BED file.}
}
\value{
A DataFrame.
}
\description{
Display select information about hub object
}
\examples{
ah = AnnotationHub()
dataInfo(ah, "BigWig")
}
